+++ pyrai/dispatcher/api.py
@@ -9,56 +9,6 @@ from pytimeparse.timeparse import timeparse
 from .structures import *
 from pyrai.helpers import to_rfc3339
 
-class Metrics:
-    """
-    Class used to specify the metrics that can be queried.
-    """
-
-    TIME = "time"
-    PASSENGERS = "passengers"
-    WAITING_REQUESTS = "waiting_requests"
-    ACTIVE_REQUESTS = "active_requests"
-    DROPPED_REQUESTS = "dropped_requests"
-    CANCELED_REQUESTS = "canceled_requests"
-    TOTAL_REQUESTS = "total_requests"
-    ASSIGNED_VEHICLES = "assigned_vehicles"
-    IDLE_VEHICLES = "idle_vehicles"
-    REBALANCING_VEHICLES = "rebalancing_vehicles"
-    OFFLINE_VEHICLES = "offline_vehicles"
-    AVG_WAIT = "avg_wait"
-    AVG_DELAY = "avg_delay"
-    AVG_OCCUPANCY = "avg_occupancy"
-    SERVICE_RATE = "service_rate"
-    QUERY = """
-    {{
-    live_fleets(
-        api_key: "{api_key}"
-        fleet_key: "{fleet_key}"
-    ) {{
-        metrics (
-        start: "{start_time}"
-        end: "{end_time}"
-        ) {{
-            time
-            passengers
-            waiting_requests
-            active_requests
-            dropped_requests
-            canceled_requests
-            total_requests
-            assigned_vehicles
-            idle_vehicles
-            rebalancing_vehicles
-            offline_vehicles
-            avg_wait
-            avg_delay
-            avg_occupancy
-            service_rate
-            }}
-        }}
-    }}
-    """
-
 class Pyrai(object):
     """
     Class used to connect to API with API key (no fleet key).
@@ -322,76 +272,8 @@ class Fleet(object):
         else:
             raise StatusError(resp=resp)
 
-    from .endpoints import make_vehicle_online
-        
-    from .endpoints import make_vehicle_offline
-    
-    from .endpoints import update_vehicle
-
-    def remove_vehicle(self, vid, location):
-        """
-        Attempts to remove a vehicle.
-
-        Args:
-            vid (int): The unique vehicle ID.
-            location (Location): The vehicle Location.
-
-        Returns:
-            list[int]: If vehicle is successfully removed, returns
-                a list of IDs of passengers of the vehicle.
-
-        Raises:
-            StatusError: If unsuccessful.
-        """
-
-        if datetime.datetime.now() > self.end_time:
-            self.end_time = datetime.datetime.now()
-        
-        url = self.build_url(Endpoints.REMOVE_VEHICLE)
-        payload = {
-            'location': location.todict(),
-            'id': vid,
-            'user_key': self.user_key.todict()
-        }
-        r = requests.post(url, data = json.dumps(payload))
-        resp = r.json()
-
-        if r.status_code == 200:
-            if resp is not None:
-                return resp.get('req_ids')
-            else:
-                return []
-        else:
-            raise StatusError(resp = resp)
-            
-    # need to update to use GraphQL but the endpoint doesn't work as expected
-    def get_vehicle_info(self, vid):
-        """
-        Attempts to get vehicle info.
-
-        Args:
-            vid (int): The unique vehicle ID.
-
-        Raises:
-            StatusError: If unsuccessful.
-
-        Returns:
-            Vehicle: If succesful, returns a vehicle object corresponding to the
-                vehicle with ID vid.
-        """
-        url = self.build_url(Endpoints.GET_VEHICLE_INFO)
-        params = {
-            'api_key': self.api_key,
-            'fleet_key': self.fleet_key,
-            'id': vid
-        }
-        r = requests.get(url, params = params)
-        resp = r.json()
-
-        if r.status_code == 200:
-            return Vehicle.fromdict(self, resp)
-        else:
-            raise StatusError(resp=resp) 
+    from pyrai.dispatcher.api_calls.fleet import make_vehicle_online, make_vehicle_offline, update_vehicle, remove_vehicle, get_vehicle_info
+   
     
     def add_request(self, rid, pickup, dropoff, load, request_time=None):
         """
@@ -658,317 +540,3 @@ class Fleet(object):
                         mode='lines+markers',
                         name=metric))
         return figure
-
-
-class UserKey(object):
-    """
-    Class for representing user keys.
-
-    Attributes:
-        api_key (string): The API key.
-        fleet_key (string): The fleet key.
-    """
-    def __init__(self, api_key, fleet_key):
-        """
-        Initializes a UserKey object.
-
-        Args:
-            api_key (string): The API key.
-            fleet_key (string): The fleet key.
-        """
-        self.api_key = api_key
-        self.fleet_key = fleet_key
-    
-    def todict(self):
-        """
-        Converts the UserKey object to a python dictionary.
-
-        Returns:
-            dict: A dictionary representation of self.
-        """
-        return {
-            'api_key': self.api_key,
-            'fleet_key': self.fleet_key
-        }
-
-    def __str__(self):
-        return str(self.todict())
-
-
-class Request(object):
-    """
-    Class for representing requests.
-
-    Attributes:
-        fleet (Fleet): The fleet that the request is a part of.
-        pickup (Location): The pickup location.
-        dropoff (Location): The dropoff location.
-        request_time (datetime.datetime): The request time.
-        req_id (int): The request ID.
-        veh_id (int): The ID of the Vehicle corresponding to this request. 
-            -1 if unassigned.
-        load (int): The load (number of passengers) in this request.
-        assigned (boolean): True if assigned, false if not.
-    """
-    def __init__(self, fleet, pickup, dropoff, request_time, req_id, veh_id, load, assigned):
-        """
-        Initializes a new Request Object.
-
-        Args:
-            fleet (Fleet): The fleet that the request is a part of.
-            pickup (Location): The pickup location.
-            dropoff (Location): The dropoff location.
-            request_time (datetime.datetime): The request time.
-            req_id (int): The request ID.
-            veh_id (int): The ID of the Vehicle corresponding to this request. 
-                -1 if unassigned.
-            load (int): The load (number of passengers) in this request.
-            assigned (boolean): True if assigned, false if not.
-        """
-
-        self.fleet = fleet
-        self.pickup = pickup
-        self.dropoff = dropoff
-        self.request_time = request_time
-        self.req_id = req_id
-        self.veh_id = veh_id
-        self.load = load
-        self.assigned = assigned
-
-    @staticmethod
-    def fromdict(fleet, d):
-        """
-        Converts a python dict into a Request object.
-
-        Args:
-            fleet (Fleet): The fleet the request is part of.
-            d (dict): The dictionary with the request parameters.
-
-        Returns:
-            Request: A request initialized with the parameters defined by d.
-        """
-        return Request(
-            fleet,
-            Location.fromdict(d.get('pickup')),
-            Location.fromdict(d.get('dropoff')),
-            isoparse(d.get('request_time')),
-            d.get('req_id'),
-            d.get('veh_id'),
-            d.get('load'),
-            d.get('assigned')
-        )
-    
-    def todict(self):
-        """
-        Converts a Request to a python dict.
-
-        Returns:
-            dict: A dictionary representation of self.
-        """
-        return {
-            'fleet': self.fleet.todict(),
-            'pickup': self.pickup.todict(),
-            'dropoff': self.dropoff.todict(),
-            'request_time': to_rfc3339(self.request_time),
-            'req_id': self.req_id,
-            'veh_id': self.veh_id,
-            'load': self.load,
-            'assigned': self.assigned
-        }
-
-    def cancel(self, event_time=None):
-        """
-        Cancels a request.
-
-        Args:
-            event_time (datetime.datetime, optional): The event time. 
-                Set to datetime.datetime.now() if not provided. Defaults to None.
-
-        Returns:
-            Status Response: If successful.
-
-        Raises:
-            StatusError: If unsuccessful.
-        """
-        if event_time is None:
-            event_time = datetime.datetime.now()
-
-        return self.fleet.cancel_request(self.req_id, event_time)
-
-    def __str__(self):
-        return str(self.todict())
-
-    def __repr__(self):
-        return str(self.todict())
-
-class VehicleEvent:
-    """
-    Describes the current event for a vehicle. 
-    `PICKUP` occurs when the vehicle has picked up a request. 
-    `DROPOFF` occurs when the vehicle has dropped of a request. 
-    `PROGRESS` should be set when the vehicle is moving to service a request, 
-    either picking up or dropping off. 
-    The vehicle should be marked as `UNASSIGNED` when it is is not 
-    assigned to any requests.
-    """
-    PICKUP = "pickup"
-    DROPOFF = "dropoff"
-    PROGRESS = "progress"
-    UNASSIGNED = "unassigned"
-
-class Notification(object):
-    """
-    Class for representing notifications.
-
-    Attributes:
-        message (str): The notification message.
-        data (NotificationData): the notification data.
-    """
-    def __init__(self, message, data):
-        """
-        Initializes a Notification object.
-
-        Args:
-            message (str): The notification message.
-            data (NotificationData): the notification data.
-        """
-        self.message = message
-        self.data = data
-    
-    @staticmethod
-    def fromdict(d):
-        """
-        Converts python dictionary to Notification object.
-
-        Args:
-            d (dict): The dictionary to convert.
-
-        Returns:
-            Notification: a Notification object with attributes set by
-                the values in d.
-        """
-        return Notification(
-            d.get('message'),
-            NotificationData.fromdict(d.get('data'))
-        )
-
-    def todict(self):
-        """
-        Converts a notification object to a python dictionary.
-
-        Returns:
-            dict: A dictionary representation of self.
-        """
-        return {
-            'message': self.message,
-            'data': self.data.todict()
-        }
-
-    def __str__(self):
-        return str(self.todict())
-
-class NotificationData(object):
-    """
-    Class used to represent notification data.
-
-    Attributes:
-        veh_id (int): The vehicle ID.
-        req_id (int): The request ID.
-        waiting_duration (str): The waiting duration.
-        assigned (bool): True if assigned, false if not.
-    """
-    def __init__(self, veh_id, req_id, waiting_duration, assigned):
-        """
-        Initializes a NotificationData object.
-
-        Args:
-            veh_id (int): The vehicle ID.
-            req_id (int): The request ID.
-            waiting_duration (str): The waiting duration.
-            assigned (bool): True if assigned, false if not.
-        """
-        self.veh_id = veh_id
-        self.req_id = req_id
-        self.waiting_duration = waiting_duration
-        self.assigned = assigned
-
-    @staticmethod
-    def fromdict(d):
-        """
-        Converts a python dictionary to a NotificationData object.
-
-        Args:
-            d (dict): The dictionary to convert.
-
-        Returns:
-            NotificationData: A NotificationData object with the attributes
-                set by values in d.
-        """
-        return NotificationData(
-            d.get('veh_id'),
-            d.get('req_id'),
-            d.get('waiting_duration'),
-            d.get('assigned')
-        )
-
-    def todict(self):
-        """
-        Converts a NotificationData object to a python dictionary.
-
-        Returns:
-            dict: A dictionary representation of self.
-        """
-
-        return {
-            'veh_id': self.veh_id,
-            'req_id': self.req_id,
-            'waiting_duration': self.waiting_duration,
-            'assigned': self.assigned
-        }
-
-    def __str__(self):
-        return str(self.todict())
-
-
-class VehicleAssignments(object):
-    """
-    Class used to represent Vehicle Assignments.
-
-    Attributes:
-        vehs (list[Vehicle]): A list of vehicles in the fleet.
-        requests (list[Request]): A list of requests in the fleet.
-        notifications (list[Notification]): A list of notifications for the fleet.
-    """
-    def __init__(self, vehs=[], requests=[], notifications=[]):
-        """
-        Initializes a VehicleAssignments object.
-
-        Args:
-            vehs (list[Vehicle], optional): A list of vehicles in the fleet. Defaults to [].
-            requests (list[Request], optional): A list of requests in the fleet. 
-                Defaults to [].
-            notifications (list[Notification], optional): A list of notifications 
-                for the fleet. Defaults to [].
-        """
-        self.vehs = vehs
-        self.requests = requests
-        self.notifications = notifications
-
-    def todict(self):
-        """
-        Converts VehicleAssignments object to python dictionary.
-
-        Returns:
-            dict: A dictionary representation of self.
-        """
-        return {
-            'vehs': [v.todict() for v in self.vehs], 
-            'requests': [r.todict() for r in self.requests],
-            'notifications': [n.todict() for n in self.notifications]
-        }
-
-    def __str__(self):
-        return str(self.todict())
-
-    def __repr__(self):
-        return str(self.todict())
