+++ pyrai/dispatcher/api_calls/__init__.py
@@ -1,0 +1,6 @@
+from .fleet import *
+from .vehicle import *
+__all__ = ["make_vehicle_online", 
+"make_vehicle_offline",
+"update_vehicle"
+]
