+++ pyrai/dispatcher/api_calls/fleet/__init__.py
@@ -1,4 +1,8 @@
 from .make_vehicle_online import make_vehicle_online
 from .make_vehicle_offline import  make_vehicle_offline
 from .update_vehicle import update_vehicle
-__all__ = ["make_vehicle_online", "make_vehicle_offline"]
+from .remove_vehicle import remove_vehicle
+__all__ = ["make_vehicle_online", 
+"make_vehicle_offline",
+"update_vehicle",
+"remove_vehicle"]
