+++ pyrai/dispatcher/api_calls/fleet/make_vehicle_online.py
