+++ pyrai/dispatcher/api_calls/fleet/update_vehicle.py
