--- pyrai/dispatcher/endpoints/__init__.py
@@ -1,7 +1,0 @@
-from .fleet import make_vehicle_online
-from .fleet import  make_vehicle_offline
-from .fleet import update_vehicle
-__all__ = ["make_vehicle_online", 
-"make_vehicle_offline",
-"update_vehicle"
-]
