+++ pyrai/dispatcher/structures/__init__.py
@@ -6,6 +6,13 @@ from .fleet_params import FleetParams
 from .event import Event
 from .location import Location
 from .vehicle import Vehicle
+from .metrics import Metrics
+from .user_key import UserKey
+from .request import Request
+from .vehicle_event import VehicleEvent
+from .notification import Notification
+from .notification_data import NotificationData
+from .vehicle_assignments import VehicleAssignments
 __all__ = ["Endpoints", 
 "StatusResponse", 
 "StatusError", 
@@ -13,5 +20,12 @@ __all__ = ["Endpoints",
 "FleetParams",
 "Event",
 "Location",
-"Vehicle"
+"Vehicle",
+"Metrics",
+"UserKey",
+"Request",
+"VehicleEvent",
+"Notification",
+"NotificationData",
+"VehicleAssignments"
 ]
